Changed TicketsService to TicketService, its path from tickets to ticket, getAll to list, getOne to retrieve, getOneExtended to retrieveExtended; Updated imports to new path and class; Added docstrings to Ticket related services, model, enums, pipes
