+++ web/src/app/components/tickets/update-ticket-status/update-ticket-status.component.ts
@@ -5,8 +5,8 @@ import { Subscription } from 'rxjs';
 import { LoadingBarService } from '@ngx-loading-bar/core';
 import { NotifyService } from 'src/app/shared/handlers/notify/notify.service';
 
-import { TicketStatus } from 'src/app/shared/services/tickets/tickets.model';
-import { TicketsService } from 'src/app/shared/services/tickets/tickets.service';
+import { TicketStatus } from 'src/app/shared/services/ticket/ticket.model';
+import { TicketService } from 'src/app/shared/services/ticket/ticket.service';
 
 @Component({
   selector: 'app-update-ticket-status',
@@ -45,7 +45,7 @@ export class UpdateTicketStatusComponent implements OnInit, OnDestroy {
     private fb: FormBuilder,
     private loadingBar: LoadingBarService,
     private notifySvc: NotifyService,
-    private ticketSvc: TicketsService
+    private ticketSvc: TicketService
   ) { }
 
   ngOnInit(): void {
