+++ web/src/app/shared/handlers/pipes/ticket-priority.pipe.ts
@@ -1,6 +1,22 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { TicketPriority } from '../../services/tickets/tickets.model';
+import { TicketPriority } from '../../services/ticket/ticket.model';
 
+/**
+ * Transforms an integer to a string representative of TicketPriority
+ * @see `TicketPriority`
+ *
+ * ### Pre-defined options
+ *
+ * | Option                 | Value            | String representative |
+ * |------------------------|------------------|-----------------------|
+ * | TicketPriority.CRIT    | 1                | Critical              |
+ * | TicketPriority.HIGH    | 2                | High                  |
+ * | TicketPriority.NORMAL  | 3                | Normal                |
+ * | TicketPriority.LOW     | 4                | Low                   |
+ * | TicketPriority.VLOW    | 5                | Very low              |
+ * | Else                   | Other than above | N/A                   |
+ * 
+ */
 @Pipe({
   name: 'ticketPriority'
 })
