+++ web/src/app/shared/services/ticket/ticket.service.spec.ts
@@ -1,13 +1,13 @@
 import { TestBed } from '@angular/core/testing';
 
-import { TicketsService } from './tickets.service';
+import { TicketService } from './ticket.service';
 
-describe('TicketsService', () => {
-  let service: TicketsService;
+describe('TicketService', () => {
+  let service: TicketService;
 
   beforeEach(() => {
     TestBed.configureTestingModule({});
-    service = TestBed.inject(TicketsService);
+    service = TestBed.inject(TicketService);
   });
 
   it('should be created', () => {
