+++ web/src/app/shared/services/ticket/ticket.service.ts
@@ -11,14 +11,17 @@ import {
   Ticket, TicketExtended, 
   TicketCommentExtended,
   TicketOverview
-} from './tickets.model';
+} from './ticket.model';
 
 const BASE_URL = `${ environment.baseUrl }v1/tickets/`
 
+/**
+ * A service for Ticket related methods
+ */
 @Injectable({
   providedIn: 'root'
 })
-export class TicketsService {
+export class TicketService {
 
   // Data
   ticket: Ticket | undefined
@@ -33,6 +36,20 @@ export class TicketsService {
     private http: HttpClient
   ) { }
 
+  /**
+   * Create new ticket entry
+   *
+   * @param body - payload
+   * @param body.title - title
+   * @param body.description - description
+   * @param body.unit - [optional] related unit ID
+   * @param body.tags - [optional] related tags ID
+   * @param body.category - [optional] category
+   * @param body.assignee - [optional] assignee ID
+   * @param body.priority - [optional] priority
+   * 
+   * @returns New created ticket entry
+   */
   create(body: any): Observable<Ticket> {
     const urlTemp = `${ BASE_URL }`
     return this.http.post<Ticket>(urlTemp, body).pipe(
@@ -43,7 +60,12 @@ export class TicketsService {
     )
   }
 
-  getAll(): Observable<Ticket[]> {
+  /**
+   * Returns a list of tickets
+   *
+   * @returns List of tickets
+   */
+  list(): Observable<Ticket[]> {
     const urlTemp = `${ BASE_URL }`
     return this.http.get<Ticket[]>(urlTemp).pipe(
       tap((res: Ticket[]) => {
@@ -53,7 +75,14 @@ export class TicketsService {
     )
   }
 
-  getOneExtended(id: number): Observable<TicketExtended> {
+  /**
+   * Returns an extended ticket
+   * 
+   * @param id - ticket ID
+   *
+   * @returns An extended ticket
+   */
+  retrieveExtended(id: number): Observable<TicketExtended> {
     const urlTemp = `${ BASE_URL }${ id }/extended/`
     return this.http.get<TicketExtended>(urlTemp).pipe(
       tap((res: TicketExtended) => {
@@ -63,6 +92,21 @@ export class TicketsService {
     )
   }
 
+  /**
+   * Partial update a ticket
+   *
+   * @param id - ticket ID
+   * @param body - payload
+   * @param body.title - [optional] title
+   * @param body.description - [optional] description
+   * @param body.unit - [optional] related unit ID
+   * @param body.tags - [optional] related tags ID
+   * @param body.category - [optional] category
+   * @param body.assignee - [optional] assignee ID
+   * @param body.priority - [optional] priority
+   * 
+   * @returns Updated ticket
+   */
   patch(id: number, body: any): Observable<Ticket> {
     const urlTemp = `${ BASE_URL }${ id }/`
     return this.http.patch<Ticket>(urlTemp, body).pipe(
@@ -73,6 +117,16 @@ export class TicketsService {
     )
   }
   
+  /**
+   * Update ticket status
+   *
+   * @param id - ticket ID
+   * @param body - payload
+   * @param body.status - status to change to
+   * @param body.notes - notes
+   * 
+   * @returns Detail response message
+   */
   updateStatus(id: number, body: any): Observable<DetailResponse> {
     const urlTemp = `${ BASE_URL }${ id }/update-status/`
     return this.http.patch<DetailResponse>(urlTemp, body).pipe(
@@ -82,6 +136,13 @@ export class TicketsService {
     )
   }
 
+  /**
+   * Returns a list of extended ticket comments
+   * 
+   * @param ticketId - ticket ID
+   *
+   * @returns List of extended ticket comments
+   */
   getComments(ticketId: number): Observable<TicketCommentExtended[]> {
     const urlTemp = `${ BASE_URL }${ ticketId }/comments/`
     return this.http.get<TicketCommentExtended[]>(urlTemp).pipe(
@@ -92,6 +153,16 @@ export class TicketsService {
     )
   }
 
+  /**
+   * Post a comment
+   *
+   * @param ticketId - ticket ID
+   * @param body - payload
+   * @param body.comment - comment
+   * @param body.replyTo - [optional] comment ID to reply to
+   * 
+   * @returns Detail response message
+   */
   postComment(ticketId: number, body: any): Observable<DetailResponse> {
     const urlTemp = `${ BASE_URL }${ ticketId }/comments/add-comment/`
     return this.http.post<DetailResponse>(urlTemp, body).pipe(
@@ -101,6 +172,11 @@ export class TicketsService {
     )
   }
 
+  /**
+   * Returns an overview of ticket
+   *
+   * @returns Overview of ticket
+   */
   getOverview(): Observable<TicketOverview> {
     const urlTemp = `${ BASE_URL }overview/`
     return this.http.get<TicketOverview>(urlTemp).pipe(
@@ -111,6 +187,11 @@ export class TicketsService {
     )
   }
 
+  /**
+   * Returns an overview of ticket status
+   *
+   * @returns Overview of ticket status
+   */
   getStatusOverview(): Observable<MultiSeries[]> {
     const urlTemp = `${ BASE_URL }status-overview/`
     return this.http.get<MultiSeries[]>(urlTemp).pipe(
@@ -121,6 +202,11 @@ export class TicketsService {
     )
   }
 
+  /**
+   * Returns an overview of ticket priority
+   *
+   * @returns Overview of ticket priority
+   */
   getPriorityOverview(): Observable<SingleSeries[]> {
     const urlTemp = `${ BASE_URL }priority-overview/`
     return this.http.get<SingleSeries[]>(urlTemp).pipe(
